// src/types/arrecadacaoAlimentoTypes.ts
export type ArrecadacaoAlimentoCreateType = {
    id_alimento: string;
    data_validade: string;
    peso: string;
    id_pessoa: string;
    id_condutor: string;
    id_deposito: string;
};

export type ArrecadacaoAlimentoUpdateType = {
    id_alimento: string;
    data_validade: string;
    peso: string;
    id_pessoa: string;
    id_condutor: string;
    id_deposito: string;
};